+++ src/lib/modlist/ModListItem.svelte
@@ -1,47 +1,21 @@
 <script lang="ts">
+	import { Button } from 'bits-ui';
 	import type { Mod } from '../models';
 	import Icon from '@iconify/svelte';
-	import { isOutdated } from '$lib/util';
-	import { readFile } from '@tauri-apps/plugin-fs';
-	import { activeGame } from '$lib/stores';
+	import { createEventDispatcher } from 'svelte';
 
 	export let mod: Mod;
 	export let isSelected: boolean;
-	export let isInstalled: boolean;
-	export let showInstalledIcon: boolean;
-	export let draggable = false;
 
-	let imgSrc: string;
+	const dispatch = createEventDispatcher<{
+		install: {
+			mod: Mod;
+		};
+	}>();
 
-	$: {
-		if (mod.type === 'remote') {
-			imgSrc = mod.icon!;
-		} else {
-			if (mod.icon) {
-				imgSrc = '';
-				loadLoadIcon(mod.icon);
-			} else {
-				imgSrc = `games/${$activeGame?.id}.webp`;
-			}
-		}
-	}
-
-	$: descriptionClasses =
-		mod.enabled === false
-			? 'text-slate-500 line-through'
-			: isSelected
-				? 'text-slate-300'
-				: 'text-slate-400 group-hover:text-slate-300';
-
-	async function loadLoadIcon(path: string) {
-		try {
-			let data = await readFile(path);
-			let blob = new Blob([data], { type: 'image/png' });
-			imgSrc = URL.createObjectURL(blob);
-		} catch (e) {
-			console.error(e);
-		}
-	}
+	$: descriptionClasses = isSelected
+		? 'text-slate-300'
+		: 'text-slate-400 group-hover:text-slate-300';
 </script>
 
 <button
@@ -50,24 +24,12 @@
 		: 'border-opacity-0 hover:bg-slate-700'}"
 	data-uuid={mod.uuid}
 	on:click
-	on:dragstart
-	on:dragover
-	on:drag
-	on:dragend
-	{draggable}
 >
-	<img src={imgSrc} alt={mod.name} class="size-12 rounded-md" />
+	<img src={mod.icon} alt={mod.name} class="size-12 rounded" />
 	<div class="flex-shrink flex-grow overflow-hidden pl-3 text-left">
 		<div class="flex items-center gap-1 overflow-hidden">
-			<div
-				class="flex-shrink truncate font-semibold {mod.enabled === false
-					? 'text-slate-300 line-through'
-					: 'text-white'}"
-			>
+			<div class="flex-shrink truncate pr-1 font-semibold text-white">
 				{mod.name.replace(/_/g, ' ')}
-			</div>
-			<div class="px-1 {descriptionClasses}">
-				{mod.version ?? ''}
 			</div>
 			{#if mod.isPinned}
 				<Icon class="flex-shrink-0 text-slate-400" icon="mdi:pin" />
@@ -75,10 +37,7 @@
 			{#if mod.isDeprecated}
 				<Icon class="flex-shrink-0 text-red-500" icon="mdi:error" />
 			{/if}
-			{#if isOutdated(mod)}
-				<Icon class="flex-shrink-0 text-green-500" icon="mdi:arrow-up-circle" />
-			{/if}
-			{#if isInstalled && showInstalledIcon}
+			{#if mod.isInstalled}
 				<Icon class="flex-shrink-0 text-green-500" icon="mdi:check-circle" />
 			{/if}
 		</div>
@@ -88,8 +47,12 @@
 		</div>
 	</div>
 
-	<!-- svelte-ignore a11y-click-events-have-key-events -->
-	<div class="contents" on:click={(evt) => evt.stopPropagation()} role="none">
-		<slot {isInstalled} />
-	</div>
+	{#if !mod.isInstalled}
+		<Button.Root
+			class="ml-2 mr-0.5 mt-0.5 hidden rounded-lg bg-green-600 p-2.5 align-middle text-2xl text-white hover:bg-green-500 group-hover:inline"
+			on:click={() => dispatch('install', { mod })}
+		>
+			<Icon icon="mdi:download" />
+		</Button.Root>
+	{/if}
 </button>